+++ NEWS.md
@@ -5,6 +5,8 @@
 * Recovery from defoliation depends on relative growth rate of buds 
 * Water potential at turgor loss point as estimation of Gsw_P50_Baldocchi
 * SpParamsMED now is filled with package traits4models
+* New output variable: CrownFoliageCompleteness
+* Rhizosphere conductance estimation with respect to resistance percent up to water potential related stomatal closure 
 
 # medfate 5.0.0
 * Partial rhizosphere overlap set to default